Fixed issues with systemcall implementations

+++ src/unistd.c
@@ -2,5 +2,5 @@
 #include "internals/syscall.h"
 
 void exit(int status){
-	__syscall1(60,status);
+	__syscall1(60,(long)status);
 }
